# contracts/platform-admin-openapi.yaml
openapi: 3.1.0
info:
  title: Merchant Platform Configuration Admin API
  version: 1.0.0
  description: >-
    Internal, TLS-only configuration control plane. Browser clients MUST use the
    same-origin admin BFF. Configuration values are immutable; every production
    change follows draft, approval, scheduling and fenced activation.
servers:
  - url: https://platform-admin.internal
security:
  - mutualTLS: []
    platformAssertion: []
...
paths:
  /internal/platform-admin/v1/changes:
    get:
      tags: [Changes]
      operationId: listPlatformChanges
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - $ref: '#/components/parameters/Kind'
        - {name: status, in: query, schema: {$ref: '#/components/schemas/ChangeStatus'}}
        - {name: cursor, in: query, schema: {type: string, format: uuid}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        '200': {description: Scoped change requests, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePage'}}}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Changes]
      operationId: createPlatformChangeDraft
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateChange'}}}}
      responses:
        '201': {description: Immutable draft version, content: {application/json: {schema: {$ref: '#/components/schemas/ChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/changes/{id}:
    get:
      tags: [Changes]
      operationId: getPlatformChange
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}]
      responses:
        '200': {description: Change request, content: {application/json: {schema: {$ref: '#/components/schemas/ChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /internal/platform-admin/v1/changes/{id}/schedule:
    post:
      tags: [Changes]
      operationId: schedulePlatformChange
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ScheduleInput'}}}}
      responses:
        '200': {description: Scheduled change, content: {application/json: {schema: {$ref: '#/components/schemas/ChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/changes/{id}/activate:
    post:
      tags: [Changes]
      operationId: manuallyActivateDuePlatformChange
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ActivateInput'}}}}
      responses:
        '200': {description: Activated immutable snapshot, content: {application/json: {schema: {$ref: '#/components/schemas/Snapshot'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/rollbacks:
    post:
      tags: [Changes]
      operationId: draftPlatformRollback
      description: Creates a new draft from a historic snapshot; it never rewrites history.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RollbackInput'}}}}
      responses:
        '201': {description: New rollback draft, content: {application/json: {schema: {$ref: '#/components/schemas/ChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/snapshots:
    get:
      tags: [Snapshots]
      operationId: listPlatformSnapshots
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - $ref: '#/components/parameters/Kind'
        - {name: logical_key, in: query, schema: {type: string}}
        - {name: cursor, in: query, schema: {type: string, format: uuid}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        '200': {description: Immutable snapshots, content: {application/json: {schema: {$ref: '#/components/schemas/SnapshotPage'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/emergency-pauses:
    post:
      tags: [Emergency]
      operationId: recordEmergencyPauseOrResume
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PauseInput'}}}}
      responses:
        '200': {description: Separately audited operational pause event}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    mutualTLS: {type: mutualTLS, description: Internal workload TLS identity enforced by the deployment boundary.}
    platformAssertion: {type: http, scheme: bearer, bearerFormat: 'PlatformAdmin compact HMAC assertion bound to method/path/query/body/audience/nonce'}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    Tenant: {name: tenant_id, in: query, required: false, description: Omit only for a globally authorized resource, schema: {type: string, format: uuid}}
...
    Kind: {name: kind, in: query, schema: {$ref: '#/components/schemas/ConfigKind'}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 255}}
...
  responses:
    Problem: {description: Fail-closed error, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    ConfigKind:
      type: string
      enum: [tenant, merchant_environment, chain, asset_contract, wallet_pool, rpc_provider, rate_source, rate_policy, finality_policy, matching_policy, quota, notification_channel, feature_flag, maintenance_window]
    ChangeStatus: {type: string, enum: [draft, approval_requested, approved, rejected, scheduled, active, superseded, cancelled]}
    CreateChange:
      type: object
      additionalProperties: false
      required: [kind, logical_key, based_on_version, payload, reason]
      properties:
        tenant_id: {type: string, format: uuid}
        kind: {$ref: '#/components/schemas/ConfigKind'}
        logical_key: {type: string, pattern: '^[a-z0-9][a-z0-9._:/-]{0,127}$'}
        based_on_version: {type: integer, format: int64, minimum: 0}
        payload: {type: object, description: Secret-free configuration. Exact monetary quantities MUST be base-10 strings.}
        reason: {type: string, minLength: 3, maxLength: 1000}
    ChangeRequest:
      allOf:
        - $ref: '#/components/schemas/CreateChange'
        - type: object
          required: [id, version, payload_hash, status, requested_by, created_at, updated_at, row_version]
          properties:
            id: {type: string, format: uuid}
            version: {type: integer, format: int64, minimum: 1}
            payload_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
            status: {$ref: '#/components/schemas/ChangeStatus'}
            requested_by: {type: string, format: uuid}
            approved_by: {type: string, format: uuid}
            rejected_by: {type: string, format: uuid}
            rollback_of_snapshot_id: {type: string, format: uuid}
            scheduled_for: {type: string, format: date-time}
            activated_at: {type: string, format: date-time}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
            row_version: {type: integer, format: int64, minimum: 1}
    Snapshot:
      type: object
      required: [id, change_request_id, kind, logical_key, version, payload, payload_hash, activated_at, fence_token]
      properties:
        id: {type: string, format: uuid}
        tenant_id: {type: string, format: uuid}
        change_request_id: {type: string, format: uuid}
        kind: {$ref: '#/components/schemas/ConfigKind'}
        logical_key: {type: string}
        version: {type: integer, format: int64}
        payload: {type: object}
        payload_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        rollback_of_snapshot_id: {type: string, format: uuid}
        activated_at: {type: string, format: date-time}
        fence_token: {type: integer, format: int64, minimum: 1}
    DecisionInput:
      type: object
      additionalProperties: false
      required: [expected_row_version, reason]
      properties: {expected_row_version: {type: integer, format: int64, minimum: 1}, reason: {type: string, minLength: 3, maxLength: 1000}}
    ScheduleInput:
      allOf:
        - $ref: '#/components/schemas/DecisionInput'
        - type: object
          required: [activate_at]
          properties: {activate_at: {type: string, format: date-time}}
    ActivateInput:
      allOf:
        - $ref: '#/components/schemas/DecisionInput'
        - type: object
          required: [expected_fence_token]
          properties: {expected_fence_token: {type: integer, format: int64, minimum: 0}}
    RollbackInput:
      type: object
      additionalProperties: false
      required: [snapshot_id, reason]
      properties: {tenant_id: {type: string, format: uuid}, snapshot_id: {type: string, format: uuid}, reason: {type: string, minLength: 3, maxLength: 1000}}
    PauseInput:
      type: object
      additionalProperties: false
      required: [kind, logical_key, action, reason]
      properties: {tenant_id: {type: string, format: uuid}, kind: {$ref: '#/components/schemas/ConfigKind'}, logical_key: {type: string}, action: {type: string, enum: [pause, resume]}, reason: {type: string, minLength: 8, maxLength: 1000}}
...
    ChangePage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/ChangeRequest'}}, next_cursor: {type: string, format: uuid}}}
    SnapshotPage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Snapshot'}}, next_cursor: {type: string, format: uuid}}}
    Problem: {type: object, required: [type, title, status, detail], properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, detail: {type: string}}}